+++ Tests/MapboxMapsTests/Location/InterpolatedLocationProducerTests.swift
@@ -84,6 +84,33 @@ final class InterpolatedLocationProducerTests: XCTestCase {
         XCTAssertEqual(returnedCancelable.cancelStub.invocations.count, 1)
     }
 
+    func testStartMulticastingPuckLocationIfEnabled() {
+        interpolatedLocationProducer.isEnabled = false
+        interpolatedLocationProducer.addPuckLocationConsumer(MockPuckLocationConsumer())
+        XCTAssertEqual(observableInterpolatedLocation.observeStub.invocations.count, 0)
+
+        interpolatedLocationProducer.isEnabled = true
+        let invocationCount = observableInterpolatedLocation.observeStub.invocations.count
+        XCTAssertEqual(invocationCount, 1)
+
+        interpolatedLocationProducer.addPuckLocationConsumer(MockPuckLocationConsumer())
+        XCTAssertEqual(observableInterpolatedLocation.observeStub.invocations.count, invocationCount)
+    }
+
+    func testPuckLocationConsumerGetNotified() {
+        let consumer = MockPuckLocationConsumer()
+        let location = InterpolatedLocation.random()
+
+        observableInterpolatedLocation.observeStub.defaultSideEffect = {
+            _ = $0.parameters(location)
+        }
+        interpolatedLocationProducer.addPuckLocationConsumer(consumer)
+
+        let receivedLocation = consumer.locationUpdateStub.invocations.first?.parameters
+        XCTAssertEqual(receivedLocation?.coordinate, location.coordinate)
+        XCTAssertEqual(receivedLocation?.accuracyAuthorization, location.accuracyAuthorization)
+    }
+
     func testParticipateBeforeInitialLocationDelivery() {
         interpolatedLocationProducer.participate()